Tiny fixes for styles and structure.
