+++ src/App.scss
@@ -936,6 +936,7 @@ body {
 
         .vitalValues {
           margin-top: 100px;
+          margin-bottom: 100px;
           padding: 0 15px;
           box-sizing: border-box;
