+++ src/InstructionComponents/FirstReport.js
@@ -1,104 +1,104 @@
 import React from "react";
 
-function Report({ personData }) {
+function Report({ controlData }) {
   const red = "#ad3e2f";
   const orange = "#db5400";
   const lightOrange = "#f0a171";
   const green = "#377d4f";
 
   const hengitystaajuusStyle = () => {
-    if (personData.Hengitystaajuus <= 8 || personData.Hengitystaajuus >= 25) {
+    if (controlData.Hengitystaajuus <= 8 || controlData.Hengitystaajuus >= 25) {
       return red;
     } else if (
-      personData.Hengitystaajuus >= 9 &&
-      personData.Hengitystaajuus <= 11
+      controlData.Hengitystaajuus >= 9 &&
+      controlData.Hengitystaajuus <= 11
     ) {
       return lightOrange;
     } else if (
-      personData.Hengitystaajuus >= 12 &&
-      personData.Hengitystaajuus <= 20
+      controlData.Hengitystaajuus >= 12 &&
+      controlData.Hengitystaajuus <= 20
     ) {
       return green;
     } else if (
-      personData.Hengitystaajuus >= 21 &&
-      personData.Hengitystaajuus <= 24
+      controlData.Hengitystaajuus >= 21 &&
+      controlData.Hengitystaajuus <= 24
     ) {
       return orange;
     }
   };
 
   const happisaturaatioStyle = () => {
-    if (personData.Happisaturaatio <= 91) {
+    if (controlData.Happisaturaatio <= 91) {
       return red;
     } else if (
-      personData.Happisaturaatio >= 92 &&
-      personData.Happisaturaatio <= 93
+      controlData.Happisaturaatio >= 92 &&
+      controlData.Happisaturaatio <= 93
     ) {
       return orange;
     } else if (
-      personData.Happisaturaatio >= 94 &&
-      personData.Happisaturaatio <= 95
+      controlData.Happisaturaatio >= 94 &&
+      controlData.Happisaturaatio <= 95
     ) {
       return lightOrange;
-    } else if (personData.Happisaturaatio >= 96) {
+    } else if (controlData.Happisaturaatio >= 96) {
       return green;
     }
   };
 
   const systolinenVerenpaineStyle = () => {
     if (
-      personData["Systolinen verenpaine"] <= 90 ||
-      personData["Systolinen verenpaine"] >= 220
+      controlData["Systolinen verenpaine"] <= 90 ||
+      controlData["Systolinen verenpaine"] >= 220
     ) {
       return red;
     } else if (
-      personData["Systolinen verenpaine"] >= 91 &&
-      personData["Systolinen verenpaine"] <= 100
+      controlData["Systolinen verenpaine"] >= 91 &&
+      controlData["Systolinen verenpaine"] <= 100
     ) {
       return orange;
     } else if (
-      personData["Systolinen verenpaine"] >= 101 &&
-      personData["Systolinen verenpaine"] <= 110
+      controlData["Systolinen verenpaine"] >= 101 &&
+      controlData["Systolinen verenpaine"] <= 110
     ) {
       return lightOrange;
     } else if (
-      personData["Systolinen verenpaine"] >= 111 &&
-      personData["Systolinen verenpaine"] <= 219
+      controlData["Systolinen verenpaine"] >= 111 &&
+      controlData["Systolinen verenpaine"] <= 219
     ) {
       return green;
     }
   };
 
   const SyketaajuusStyle = () => {
-    if (personData.Syketaajuus <= 40 || personData.Syketaajuus >= 131) {
+    if (controlData.Syketaajuus <= 40 || controlData.Syketaajuus >= 131) {
       return red;
-    } else if (personData.Syketaajuus >= 111 && personData.Syketaajuus <= 130) {
+    } else if (controlData.Syketaajuus >= 111 && controlData.Syketaajuus <= 130) {
       return orange;
     } else if (
-      (personData.Syketaajuus >= 41 && personData.Syketaajuus <= 50) ||
-      (personData.Syketaajuus >= 91 && personData.Syketaajuus <= 110)
+      (controlData.Syketaajuus >= 41 && controlData.Syketaajuus <= 50) ||
+      (controlData.Syketaajuus >= 91 && controlData.Syketaajuus <= 110)
     ) {
       return lightOrange;
-    } else if (personData.Syketaajuus >= 51 && personData.Syketaajuus <= 90) {
+    } else if (controlData.Syketaajuus >= 51 && controlData.Syketaajuus <= 90) {
       return green;
     }
   };
 
   const lampotilaStyle = () => {
-    if (personData["Mittaa lämpötila"] <= 35) {
+    if (controlData["Mittaa lämpötila"] <= 35) {
       return red;
-    } else if (personData["Mittaa lämpötila"] >= 39.1) {
+    } else if (controlData["Mittaa lämpötila"] >= 39.1) {
       return orange;
     } else if (
-      (personData["Mittaa lämpötila"] >= 35.1 &&
-        personData["Mittaa lämpötila"] <= 36) ||
-      (personData["Mittaa lämpötila"] >= 38.1 &&
-        personData["Mittaa lämpötila"] <= 39)
+      (controlData["Mittaa lämpötila"] >= 35.1 &&
+        controlData["Mittaa lämpötila"] <= 36) ||
+      (controlData["Mittaa lämpötila"] >= 38.1 &&
+        controlData["Mittaa lämpötila"] <= 39)
     ) {
       return lightOrange;
     } else if (
-      personData["Mittaa lämpötila"] >= 36.1 &&
-      personData["Mittaa lämpötila"] <= 38
+      controlData["Mittaa lämpötila"] >= 36.1 &&
+      controlData["Mittaa lämpötila"] <= 38
     ) {
       return green;
     }
@@ -106,25 +106,25 @@ function Report({ personData }) {
 
   const verensokeriStyle = () => {
     if (
-      personData["Mittaa verensokeri:"] <= 4 ||
-      personData["Mittaa verensokeri:"] >= 25
+      controlData["Mittaa verensokeri:"] <= 4 ||
+      controlData["Mittaa verensokeri:"] >= 25
     ) {
       return red;
     } else if (
-      (personData["Mittaa verensokeri:"] >= 5 &&
-        personData["Mittaa verensokeri:"] <= 6) ||
-      (personData["Mittaa verensokeri:"] >= 20 &&
-        personData["Mittaa verensokeri:"] <= 24)
+      (controlData["Mittaa verensokeri:"] >= 5 &&
+        controlData["Mittaa verensokeri:"] <= 6) ||
+      (controlData["Mittaa verensokeri:"] >= 20 &&
+        controlData["Mittaa verensokeri:"] <= 24)
     ) {
       return orange;
     } else if (
-      personData["Mittaa verensokeri:"] >= 15 &&
-      personData["Mittaa verensokeri:"] <= 20
+      controlData["Mittaa verensokeri:"] >= 15 &&
+      controlData["Mittaa verensokeri:"] <= 20
     ) {
       return lightOrange;
     } else if (
-      personData["Mittaa verensokeri:"] >= 7 &&
-      personData["Mittaa verensokeri:"] <= 14
+      controlData["Mittaa verensokeri:"] >= 7 &&
+      controlData["Mittaa verensokeri:"] <= 14
     ) {
       return green;
     }
@@ -132,14 +132,18 @@ function Report({ personData }) {
 
   return (
     <div className="report-container">
+      <p style={{color: "gray"}}>*Time stamp*</p>
+
       <h2>Raportti:</h2>
 
       <div className="reportElement">
+
         <h3>Hengitystie:</h3>
+
         <div className="flex">
           <p className="subject">{"Onko hengitystie auki?"}</p>
           <div className="answer">
-            {personData["Onko hengitystie auki?"] ? (
+            {controlData["Onko hengitystie auki?"] ? (
               <p style={{ color: green }}>Kyllä</p>
             ) : (
               <p style={{ color: red }}>Ei</p>
@@ -149,7 +153,7 @@ function Report({ personData }) {
         <div className="flex">
           <p className="subject">{"Onko ilmatie estettä?"}</p>
           <div className="answer">
-            {personData["Onko ilmatie estettä?"] ? (
+            {controlData["Onko ilmatie estettä?"] ? (
               <p style={{ color: red }}>Kyllä</p>
             ) : (
               <p style={{ color: green }}>Ei</p>
@@ -166,18 +170,18 @@ function Report({ personData }) {
           <p className="subject">{"Korvin kuultavat äänet?"}</p>
           <div className="answer">
             <p style={{ color: green }}>
-              {personData["Hengitys - Normaali hengitys/ei ääniä"]
+              {controlData["Hengitys - Normaali hengitys/ei ääniä"]
                 ? "Normaali"
                 : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Hengitys - Vinkuna"] ? "Vinkuna" : null}
+              {controlData["Hengitys - Vinkuna"] ? "Vinkuna" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Hengitys - Korina"] ? "Korina" : null}
+              {controlData["Hengitys - Korina"] ? "Korina" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Hengitys - Rohina"] ? "Rohina" : null}
+              {controlData["Hengitys - Rohina"] ? "Rohina" : null}
             </p>
           </div>
         </div>
@@ -190,7 +194,7 @@ function Report({ personData }) {
         <div className="flex">
           <p className="subject">{"Syke"}</p>
           <div className="answer">
-            {personData["Tarkista syke:"] ? (
+            {controlData["Tarkista syke:"] ? (
               <p style={{ color: green }}>Säännöllinen</p>
             ) : (
               <p style={{ color: red }}>Epäsäännöllinen</p>
@@ -200,7 +204,7 @@ function Report({ personData }) {
         <div className="flex">
           <p className="subject">{"Tuntuuko paikallista lämpöeroa?"}</p>
           <div className="answer">
-            {personData["Tuntuuko lämpörajoja raajoissa:"] ? (
+            {controlData["Tuntuuko lämpörajoja raajoissa:"] ? (
               <p style={{ color: red }}>Kyllä</p>
             ) : (
               <p style={{ color: green }}>Ei</p>
@@ -215,9 +219,9 @@ function Report({ personData }) {
         <h3>Tajunta:</h3>
         <div className="flex">
           <p style={{ width: "100%", fontWeight: "200" }} className="answer">
-            {personData["Tajunta - Ei poikkeavia löydöksiä."]
+            {controlData["Tajunta - Ei poikkeavia löydöksiä."]
               ? "Ei poikkeavia löydöksiä."
-              : personData.Tajunta}
+              : controlData.Tajunta}
           </p>
         </div>
       </div>
@@ -230,33 +234,33 @@ function Report({ personData }) {
           <p className="subject">{"Miltä iho tuntuu?"}</p>
           <div className="answer">
             <p style={{ color: green }}>
-              {personData["Iho, paljastaminen - Normaali"] ? "Normaali" : null}
+              {controlData["Iho, paljastaminen - Normaali"] ? "Normaali" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Iho, paljastaminen - Kuiva"] ? "Kuiva" : null}
+              {controlData["Iho, paljastaminen - Kuiva"] ? "Kuiva" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Iho, paljastaminen - Kostea"] ? "Kostea" : null}
+              {controlData["Iho, paljastaminen - Kostea"] ? "Kostea" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Iho, paljastaminen - Kylmä"] ? "Kylmä" : null}
+              {controlData["Iho, paljastaminen - Kylmä"] ? "Kylmä" : null}
             </p>
             <p style={{ color: red }}>
-              {personData["Iho, paljastaminen - Kuuma"] ? "Kuuma" : null}
+              {controlData["Iho, paljastaminen - Kuuma"] ? "Kuuma" : null}
             </p>
           </div>
         </div>
         <div className="flex">
           <p className="subject">Onko iho muutoksia?</p>
           <div className="answer">
-            {personData["Iho, paljastaminen - Ei poikkeavia löydöksiä"] ? (
+            {controlData["Iho, paljastaminen - Ei poikkeavia löydöksiä"] ? (
               <p style={{ color: green }}>Ei</p>
             ) : (
               <p style={{ color: red }}>Kyllä</p>
             )}
           </div>
         </div>
-        {personData.Iho ? (
+        {controlData.Iho ? (
           <p
             style={{
               fontWeight: "200",
@@ -265,7 +269,7 @@ function Report({ personData }) {
             }}
             className="answer"
           >
-            {personData.Iho}
+            {controlData.Iho}
           </p>
         ) : null}
       </div>
@@ -277,13 +281,13 @@ function Report({ personData }) {
         <div className="flex">
           <p className="vital-subject">{"hengitystaajuus:"}</p>
           <p className="vital-answer" style={{ color: hengitystaajuusStyle() }}>
-            {personData.Hengitystaajuus}
+            {controlData.Hengitystaajuus}
           </p>
         </div>
         <div className="flex">
           <p className="vital-subject">{"Happisaturaatio:"}</p>
           <p className="vital-answer" style={{ color: happisaturaatioStyle() }}>
-            {personData.Happisaturaatio}
+            {controlData.Happisaturaatio}
           </p>
         </div>
         <div className="flex">
@@ -292,19 +296,19 @@ function Report({ personData }) {
             className="vital-answer"
             style={{ color: systolinenVerenpaineStyle() }}
           >
-            {personData["Systolinen verenpaine"]}
+            {controlData["Systolinen verenpaine"]}
           </p>
         </div>
         <div className="flex">
           <p className="vital-subject">{"Syketaajuus:"}</p>
           <p className="vital-answer" style={{ color: SyketaajuusStyle() }}>
-            {personData.Syketaajuus}
+            {controlData.Syketaajuus}
           </p>
         </div>
         <div className="flex">
           <p className="vital-subject">{"Tajunnan taso:"}</p>
           <div className="vital-answer">
-            {personData["Tajunnan taso"] ? (
+            {controlData["Tajunnan taso"] ? (
               <p style={{ color: green }}>Normaali</p>
             ) : (
               <p style={{ color: red }}>Poikkeava</p>
@@ -314,15 +318,15 @@ function Report({ personData }) {
         <div className="flex">
           <p className="vital-subject">{"Lämpötila:"}</p>
           <p className="vital-answer" style={{ color: lampotilaStyle() }}>
-            {personData["Mittaa lämpötila"]}
+            {controlData["Mittaa lämpötila"]}
           </p>
         </div>
         <div className="flex">
           <p className="vital-subject">{"Verensokeri:"}</p>
           <p className="vital-answer" style={{ color: verensokeriStyle() }}>
-            {personData["Mittaa verensokeri:"]}
+            {controlData["Mittaa verensokeri:"]}
           </p>
-        </div>
+        </div> <hr></hr>
       </div>
     </div>
   );
